delete : delete unreusable component
